backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
requests==2.31.0
gunicorn==21.2.0
bcrypt==4.0.1
